feat: kafka [S13P31A303-216] ### Kafka - application.yml에 Kafka 기본 설정 추가 - Producer / Consumer 기본 동작 검증 완료 - BidEventMessage 기반 입찰 이벤트 송수신 구현
### Redis (입찰 기능)
- bidRedisTemplate 설정 추가
- 실시간 입찰 점수 저장용 Sorted Set(ZSET) 구성
- 유저별 최초 1회 입찰 제한 로직 분리 (BidTryCountRepository)
- 진행 중 경매 여부 체크용 Redis Value 구성 (AuctionInProgressRepository)

### Ranking & Persistence
- Redis ZSET 기반 실시간 입찰 순위 관리
- 종료된 경매의 Redis 데이터를 DB로 저장하는 로직 구현
- 종료 후 Redis 키(bid, bid-try, auction-in-progress) 정리

### Auction 스케줄링
- Auction 엔티티에 현재 상태(progress/finish) 필드 추가
- 시작해야 할 경매 DB에서 조회 후 Redis에 등록 (스케줄링)
- 종료된 경매는 매일 12시에 flush 후 DB 저장 & finish 처리

### Bid 조회
- 특정 매물의 경매 조회
- 모든 경매 내역 조회

### 공통 코드
- DataSerializer 도입으로 JSON 직렬화/역직렬화 공통화

+++ backend/src/main/java/ssafy/a303/backend/auction/repository/BidRankRepository.java
@@ -23,7 +23,7 @@ public class BidRankRepository {
         redis.opsForZSet().add(
                 generateKey(message.auctionSeq()),
                 json,
-                message.amount()
+                message.amount().doubleValue()
         );
     }
